+++ assets/javascripts/kitten/helpers/utils/lazy-observer.js
@@ -1,7 +1,0 @@
-import { IntersectionObserverClass } from './intersection-observer'
-
-const LazyObserver = new IntersectionObserverClass({
-  rootMargin: '100px 0px',
-})
-
-export default LazyObserver